misc/HtmlTestSuite/results/test-0112.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Priority Hints - Image element</title>
<meta href="mailto:domfarolino@gmail.com" name="author" title="Dominic Farolino"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

</head><body><img fetchpriority="high" id="img1" src="/images/green.png"/>
<img fetchpriority="low" id="img2" src="/images/green.png"/>
<img fetchpriority="auto" id="img3" src="/images/green.png"/>
<img fetchpriority="xyz" id="img4" src="/images/green.png"/>
<img id="img5" src="/images/green.png"/>

<script>
  test(() =&gt; {
    assert_equals(img1.fetchPriority, "high", "high fetchPriority is a valid IDL value on the image element");
    assert_equals(img2.fetchPriority, "low", "low fetchPriority is a valid IDL value on the image element");
    assert_equals(img3.fetchPriority, "auto", "auto fetchPriority is a valid IDL value on the image element");
    assert_equals(img4.fetchPriority, "auto", "invalid fetchPriority reflects as 'auto' IDL attribute on the image element");
    assert_equals(img5.fetchPriority, "auto", "missing fetchPriority reflects as 'auto' IDL attribute on the image element");
  }, "fetchpriority attribute on &lt;img&gt; elements should reflect valid IDL values");

  test(() =&gt; {
    const img = new Image();
    assert_equals(img.fetchPriority, "auto");
  }, "fetchPriority of new Image() is 'auto'");
</script>
</body></html>
